feat: gen 自动生成 gorm 模型

+++ internal/api/admin/func_createadminmenu.go
@@ -14,7 +14,7 @@ type createAdminMenuRequest struct {
 }
 
 type createAdminMenuResponse struct {
-	Id int32 `json:"id"` // 主键ID
+	Id int64 `json:"id"` // 主键ID
 }
 
 // CreateAdminMenu 提交菜单授权
@@ -53,7 +53,7 @@ func (h *handler) CreateAdminMenu() core.HandlerFunc {
 		}
 
 		createData := new(admin.CreateMenuData)
-		createData.AdminId = int32(ids[0])
+		createData.AdminId = int64(ids[0])
 		createData.Actions = req.Actions
 
 		err = h.adminService.CreateMenu(c, createData)
@@ -66,7 +66,7 @@ func (h *handler) CreateAdminMenu() core.HandlerFunc {
 			return
 		}
 
-		res.Id = int32(ids[0])
+		res.Id = int64(ids[0])
 		c.Payload(res)
 	}
 }
